requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
et_xmlfile==2.0.0
filelock==3.16.1
frozenlist==1.5.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
jiter==0.7.0
jmespath==1.0.1
lxml==5.3.0
multidict==6.1.0
mysql==0.0.3
mysqlclient==2.2.6
numpy==2.1.2
oauthlib==3.2.2
openai==1.54.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parsel==1.9.1
propcache==0.2.0
Protego==0.3.1
proto-plus==1.25.0
protobuf==5.28.3
py3dns==4.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyDispatcher==2.0.7
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
Scrapy==2.11.2
service-identity==24.2.0
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tldextract==5.1.2
tqdm==4.66.6
Twisted==24.10.0
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
validate-email-address==1
w3lib==2.2.1
wheel==0.44.0
yarl==1.17.1
zope.interface==7.1.1

trio~=0.27.0
pip~=24.3.1
PySocks~=1.7.1
pillow~=11.0.0
outcome~=1.3.0.post0
selenium~=4.27.1
email_validator~=2.2.0